requirements.txt
fastapi
uvicorn
pydantic[email]
python-jose
argon2-cffi
duckdb
nanoid
python-multipart
